+++ app/Http/Controllers/Auth/RegisterController.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace App\Http\Controllers\Auth;
+
+use App\Http\Controllers\Controller;
+use App\Models\User;
+use Illuminate\Http\Request;
+
+class RegisterController extends Controller
+{
+
+    public function index(){
+        return view('auth.register');
+    }
+
+    public function process(Request $request){
+
+        $user = $request->validate([
+            'name' => ['required'],
+            'email' => ['required', 'email'],
+            'password' => ['required'],
+        ]);
+
+        User::create($user);
+
+        return redirect()->route('login');
+    }
+}