+++ database/migrations/2024_04_19_143224_create_files_table.php
@@ -1,0 +1,37 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::create('files', function(Blueprint $table){
+            $table->id();
+            $table->foreignId('user_id')->nullable()->index();
+            $table->string('name');
+            $table->string('ext');
+            $table->string('size');
+            $table->string('path');
+            $table->string('code')->unique();
+            $table->string('description')->nullable();
+            $table->string('password')->nullable();
+            $table->string('downloaded')->default(0);
+            $table->string('viewed')->default(0);
+            $table->timestamps();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::dropIfExists('files');
+    }
+};