+++ resources/views/file.blade.php
@@ -8,40 +8,46 @@
                     <div class="card-header rounded-0 bg-secondary">
                         <h1 class="fs-3 text-white">
                             <i class="bi bi-file"></i>
-                            FileName
+                            {{ $file['name'] }}
                         </h1>
                     </div>
                     <ul class="list-group list-group-flush">
                         <li class="list-group-item">
                             <i class="bi bi-file-code"></i>
-                            file type
+                            {{ $file['ext'] }}
                         </li>
                         <li class="list-group-item">
                             <i class="bi bi-person"></i>
-                            <a class="text-decoration-none" href="user/username">user</a> on <a class="text-decoration-none" href="category/category">category</a>
+                            <a class="text-decoration-none" href="{{ url('user/'.$user['name']) }}">
+                                {{ $user['name'] }}
+                            </a>
+                            on
+                            <a class="text-decoration-none" href="category/{{ $file['ext'] }}">
+                                {{ $file['ext'] }}
+                            </a>
                         </li>
                         <li class="list-group-item">
                             <i class="bi bi-upload"></i>
-                            Uploaded: 19 may 2029
+                            Uploaded: {{ $file['created_at'] }}
                         </li>
                         <li class="list-group-item">
                             <i class="bi bi-download"></i>
-                            Downlaods: 10
+                            Downlaods: {{ $file['downloaded'] }}
                         </li>
                     </ul>
                 </div>
 
                 {{-- download button --}}
                 <div class="text-center mb-3">
-                    <a class="btn btn-primary" href="download/file">
-                        Download File
+                    <a class="btn btn-primary" href="{{ url('download/'. $file['code']) }}">
+                        Download {{ $file['name'] }}
                     </a>
                 </div>
 
                 <div class="card rounded-0 mb-3">
-                    <div class="card-body">
+                    <div class="card-body text-center">
                         <p>
-                            Download Filename by username
+                            Download {{ $file['name'] }} by {{ $user['name'] }}
                         </p>
                         <p>
                             Lorem ipsum dolor sit amet, consectetur adipisicing elit. Ab ea maxime repellendus totam vitae? Alias aliquid atque blanditiis dignissimos distinctio dolor impedit laboriosam laudantium optio quasi, repellat saepe sint voluptatem.
@@ -50,7 +56,7 @@
 
                     <ul class="list-group list-group-flush">
                         <li class="list-group-item">
-                            <input type="text" class="form-control" value="file_url">
+                            <input type="text" class="form-control w-50" value="{{ url($file['code']) }}" readonly>
                         </li>
 
                         <li class="list-group-item">