+++ resources/views/layouts/app-layout/header.blade.php
@@ -1,4 +1,4 @@
-<nav class="navbar navbar-expand-lg navbar-dark bg-primary">
+<nav class="navbar navbar-expand-lg navbar-dark bg-primary shadow">
     <div class="container-fluid">
 
         <x-app-name class="navbar-brand"/>
@@ -10,17 +10,17 @@
         <div class="collapse navbar-collapse me-auto" id="navbarNav">
             <ul class="navbar-nav ms-auto">
                 <li class="nav-item">
-                    <a class="nav-link {{ request()->is('upload') ? 'active' : '' }}" href="{{ route('upload') }}">
+                    <a class="nav-link {{ request()->is('/') ? 'active' : '' }}" href="{{ route('home') }}">
                         <i class="bi bi-upload"></i>
                         Upload
                     </a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ route('latest') }}">Latest
+                    <a class="nav-link {{ request()->is('latest') ? 'active' : '' }}" href="{{ route('latest') }}">Latest
                     </a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ route('trending') }}">Trending
+                    <a class="nav-link {{ request()->is('trending') ? 'active' : '' }}" href="{{ route('trending') }}">Trending
                     </a>
                 </li>
 
@@ -39,11 +39,11 @@
                     </li>
                 @else
                     <li class="nav-item">
-                        <a class="nav-link" href="{{ route('login') }}">Login
+                        <a class="nav-link {{ request()->is('login') ? 'active' : '' }}" href="{{ route('login') }}">Login
                         </a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="{{ route('register') }}">Register
+                        <a class="nav-link {{ request()->is('register') ? 'active' : '' }}" href="{{ route('register') }}">Register
                         </a>
                     </li>
                 @endif