+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # ChangeLog
+**17 Mei 2024**
+- refactory routes move to controller
+- add migration files
+- add model files
+- upload & download process
+- latest page
+- add php module [pharaonic/laravel-readable]
+
+**15 Mei 2024**
+- add js module [@hotwired/turbo]
 
 **12 Mei 2024**
 - add page file, download, user